chore: move Celo tests to right file
